Updated directory structure to include src/ and tests/.
