--- Microsoft.Extensions.Configuration.AppConfig/IKeyValueFormatter.cs
@@ -1,7 +1,0 @@
-﻿namespace Microsoft.Extensions.Configuration.AppConfig
-{
-    public interface IKeyValueFormatter
-    {
-        string Format(IKeyValue keyValue);
-    }
-}
