--- Microsoft.Extensions.Configuration.AppConfig/KeyValueFormatter.cs
@@ -1,10 +1,0 @@
-﻿namespace Microsoft.Extensions.Configuration.AppConfig
-{
-    class KeyValueFormatter : IKeyValueFormatter
-    {
-        public string Format(IKeyValue keyValue)
-        {
-            return keyValue.Value;
-        }
-    }
-}
