+++ examples/ConfigStoreDemo/ConfigStoreDemo.csproj
@@ -9,7 +9,7 @@
     <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0.2" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\Microsoft.Extensions.Configuration.AppConfig\Microsoft.Extensions.Configuration.AppConfig.csproj" />
+    <ProjectReference Include="..\..\src\Microsoft.Extensions.Configuration.AppConfig\Microsoft.Extensions.Configuration.AppConfig.csproj" />
   </ItemGroup>
   <ItemGroup>
     <Content Update="appsettings.json">
