+++ src/Microsoft.Extensions.Configuration.AppConfig/IKeyValue.cs
@@ -1,0 +1,14 @@
+﻿namespace Microsoft.Extensions.Configuration.AppConfig
+{
+    using System;
+
+    public interface IKeyValue
+    {
+        string Key { get; set; }
+        string Value { get; set; }
+        string ContentType { get; set; }
+        string ETag { get; set; }
+        DateTimeOffset Created { get; set; }
+        DateTimeOffset LastModified { get; set; }
+    }
+}
