+++ src/Microsoft.Extensions.Configuration.AppConfig/IKeyValueFormatter.cs
@@ -1,0 +1,7 @@
+﻿namespace Microsoft.Extensions.Configuration.AppConfig
+{
+    public interface IKeyValueFormatter
+    {
+        string Format(IKeyValue keyValue);
+    }
+}
