+++ src/Microsoft.Extensions.Configuration.AppConfig/KeyValueFormatter.cs
@@ -1,0 +1,10 @@
+﻿namespace Microsoft.Extensions.Configuration.AppConfig
+{
+    class KeyValueFormatter : IKeyValueFormatter
+    {
+        public string Format(IKeyValue keyValue)
+        {
+            return keyValue.Value;
+        }
+    }
+}
