+++ src/Microsoft.Extensions.Configuration.AppConfig/KeyValueListener.cs
@@ -1,0 +1,13 @@
+﻿namespace Microsoft.Extensions.Configuration.AppConfig
+{
+    using System;
+
+    public class KeyValueListener
+    {
+        public string Key { get; set; }
+
+        public int PollInterval { get; set; }
+
+        public Action Callback { get; set; }
+    }
+}
