+++ src/Microsoft.Extensions.Configuration.AppConfig/RemoteConfigurationOptions.cs
@@ -1,0 +1,33 @@
+﻿namespace Microsoft.Extensions.Configuration.AppConfig
+{
+    using System.Collections.Generic;
+
+    public class RemoteConfigurationOptions
+    {
+        private Dictionary<string, KeyValueListener> _changeListeners = new Dictionary<string, KeyValueListener>();
+
+        public string AcceptVersion { get; set; } = null;
+
+        public string Prefix { get; set; } = string.Empty;
+
+        public IKeyValueFormatter KeyValueFormatter { get; set; } = new KeyValueFormatter();
+
+        public IEnumerable<KeyValueListener> ChangeListeners {
+            get {
+                return _changeListeners.Values;
+            }
+        }
+
+        public RemoteConfigurationOptions Listen(string key, int pollInterval)
+        {
+            _changeListeners[key] = new KeyValueListener()
+            {
+                Key = key,
+                PollInterval = pollInterval,
+                Callback = null
+            };
+
+            return this;
+        }
+    }
+}
